Rename URLs for editing user data
**Why**:
* They were inconsistent before
* Now they are all consistently part of `manage` namespace
* It is not part of Rails conventions to name controllers `edit_xyz`.
* In Rails, controllers are named after nouns and have verb methods
* Also addresses two factor namespace issue raised at https://github.com/18F/identity-idp/pull/848#discussion_r92094515
