+++ app/views/users/emails/edit.html.slim
@@ -2,7 +2,7 @@
 
 
 h1.h3.my0 = t('headings.edit_info.email')
-= simple_form_for(@update_user_email_form, url: edit_email_path,
+= simple_form_for(@update_user_email_form, url: manage_email_path,
     html: { autocomplete: 'off', method: :patch, role: 'form' }) do |f|
   = f.error_notification
   = f.input :email, required: true
