+++ app/views/users/phones/edit.html.slim
@@ -2,7 +2,7 @@
 
 
 h1.h3.my0 = t('headings.edit_info.phone')
-= simple_form_for(@update_user_phone_form, url: edit_phone_path,
+= simple_form_for(@update_user_phone_form, url: manage_phone_path,
     html: { autocomplete: 'off', method: :put, role: 'form' }) do |f|
   = f.input :phone, as: :tel, required: true
   = f.button :submit, t('forms.buttons.submit.update'), class: 'mt2'
