+++ spec/features/users/user_edit_spec.rb
@@ -4,7 +4,7 @@ feature 'User edit' do
   scenario 'user sees error message if form is submitted without email', js: true do
     sign_in_and_2fa_user
 
-    visit edit_email_path
+    visit manage_email_path
     fill_in 'Email', with: ''
     click_button 'Update'
 
@@ -14,7 +14,7 @@ feature 'User edit' do
   scenario 'user sees error message if form is submitted without phone number', js: true do
     sign_in_and_2fa_user
 
-    visit edit_phone_path
+    visit manage_phone_path
     fill_in 'Phone', with: ''
     click_button 'Update'
