+++ spec/views/users/emails/edit.html.slim_spec.rb
@@ -1,15 +1,15 @@
 require 'rails_helper'
 
-describe 'users/edit_phone/edit.html.slim' do
+describe 'users/emails/edit.html.slim' do
   context 'user is not TOTP enabled' do
     before do
       user = build_stubbed(:user, :signed_up)
       allow(view).to receive(:current_user).and_return(user)
-      @update_user_phone_form = UpdateUserPhoneForm.new(user)
+      @update_user_email_form = UpdateUserEmailForm.new(user)
     end
 
     it 'has a localized title' do
-      expect(view).to receive(:title).with(t('titles.edit_info.phone'))
+      expect(view).to receive(:title).with(t('titles.edit_info.email'))
 
       render
     end
